fixup! Review feedback: special case for CI
More complicated than it should be: this allows a user to
check in a bogus specfile tree and use that in a CI setting.
To do this I've moved the major config settings (git repo
URL and branch) from Makefile to the make-testonly-rpms
script itself. It now has a hardcoded fedoraproject.org
URL, overrideable via environment variable or via the
presence of a special directory.

Also:
 - add -E flag to sudo invocation, and explain why
 - add emergency cleanup of workdir
 - document BASHisms
 - add sanity check for missing Source1 URL

Signed-off-by: Ed Santiago <santiago@redhat.com>